+++ http_cache_client/lib/src/client/http_cache_client_cache_utils.dart
@@ -94,6 +94,7 @@ extension _CacheClientUtils on CacheClient {
           headers: request.inner.headers,
         ),
         options: request.options,
+        requestDate: request.requestDate,
       ),
     );