+++ http_cache_client/CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 1.0.1
 - fix: Improve cache save time.
+- fix: Improve request date assignment.
 
 ## 1.0.0
 - Initial version.